Create HTML_V2.cs
There are definately some issues, now I have to make sure that the formatting (css) gets to be written before the body tag and of course 
make sure is wrapped (with <>) in the style tag. 

I want to make this more robust as I can read the date, have it sort as such and remove the ECU info as it is redundant. 

The comments are important as I write what I have done to my car. 

Overall this is shaping to be a sort of log of DTC codes that will be uploaded to a website for record keeping and view