Create Website.cs
The last iteration of the HTML class relied on creating the tags pragmatically and adding info from a txt file to the body of the webpage. 
This class now will read from an HTML template and when it encounters the <body> or <style> tags it will read from seperate templates respectively. This aim is to automate websites I create with my own style and will then update with DTC engine codes from my car.